Sp7

+++ CH_1-Introduction/1-3 CSES_4/Weird algorithms long long.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main() {
 
-	int n;
+	long int n;
 	cout << "Enter an integer: ";
 	cin >> n;
